Module M.P256

include Tezos_crypto.P256

include Tezos_crypto.S.SIGNATURE with type watermark = Stdlib.Bytes.t

module Public_key_hash : sig ... end

module Public_key : sig ... end

module Secret_key : sig ... end

type t

val pp : Stdlib.Format.formatter -> t -> unit

include Tezos_stdlib.Compare.S with type t := t

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

include Tezos_crypto.S.B58_DATA with type t := t

type t

val to_b58check : t -> string

val to_short_b58check : t -> string

val of_b58check : string -> t Tezos_error_monad.Error_monad.tzresult

val of_b58check_exn : string -> t

val of_b58check_opt : string -> t option

type Tezos_crypto.Base58.data += Data of t

val b58check_encoding : t Tezos_crypto.Base58.encoding

include Tezos_crypto.S.ENCODER with type t := t

type t

val encoding : t Data_encoding.t

val rpc_arg : t Tezos_rpc.RPC_arg.t

val zero : t

type watermark = Stdlib.Bytes.t

val sign : ?⁠watermark:watermark -> Secret_key.t -> Stdlib.Bytes.t -> t

val check : ?⁠watermark:watermark -> Public_key.t -> t -> Stdlib.Bytes.t -> bool

val generate_key : ?⁠seed:Stdlib.Bytes.t -> unit -> Public_key_hash.t * Public_key.t * Secret_key.t

val deterministic_nonce : Secret_key.t -> Stdlib.Bytes.t -> Stdlib.Bytes.t

deterministic_nonce sk msg returns a nonce that is determined by sk and msg

val deterministic_nonce_hash : Secret_key.t -> Stdlib.Bytes.t -> Stdlib.Bytes.t

deterministic_nonce_hash sk msg returns the BLAKE2b hash of a nonce that is determined by sk and msg.

In other words, Blake2b.digest (deterministic_nonce sk msg) = deterministic_nonce_hash sk msg

include Tezos_crypto.S.RAW_DATA with type t := t

type t

val size : int

val to_hex : t -> Hex.t

val of_hex : Hex.t -> t Tezos_error_monad.Error_monad.tzresult

val of_hex_opt : Hex.t -> t option

val of_hex_exn : Hex.t -> t

val to_string : t -> string

val of_string : string -> t Tezos_error_monad.Error_monad.tzresult

val of_string_opt : string -> t option

val of_string_exn : string -> t

val to_bytes : t -> Stdlib.Bytes.t

val of_bytes : Stdlib.Bytes.t -> t Tezos_error_monad.Error_monad.tzresult

val of_bytes_opt : Stdlib.Bytes.t -> t option

val of_bytes_exn : Stdlib.Bytes.t -> t

module Public_key_hash : sig ... end